Add client support for SCRAM-SHA512 auth for API keys
This commit adds the ability for our API client and midclt command to
perform SCRAM-SHA512 authentication to a TrueNAS server. This new
authentication mechanism provides replay resistance and detection of
case where the remote server is (possibly maliciously) granting us
access without actually having access to either the API key's hash or
keys derived from said hash.

The commit also adds a variety of scram-related helper functions
that will be used in TrueNAS and in CI tests. The new authentication
mechanism is only attempted if the TrueNAS server reports that it
supports the mechanism.